requirements.txt
blynk-library-python==0.2.0
network==0.1
RPi.GPIO==0.7.1